+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 ### Removed
 - Models: Moved `IndexedSet` and `IndexedSet` to Primitives.
 
+### Fixed
+- Traversal: Exception in indexer for `IndexedColorDictionary`.
+
 ## [0.10.0] - 2021-05-30
 ### Added
 - Models: `CompactSet` structure implementing `ISet<int>` with bit array as backing store.